requirements.txt
flask
requests
pandas
numpy
scikit-learn
scikit-surprise